requirements.txt
asgiref==3.7.2
channels==4.0.0
channels_redis==4.1.0
django==5.0.6
djangorestframework==3.14.0
Requests==2.31.0
sseclient_py==1.8.0
stripe==9.4.0
django-cors-headers==4.2.0
daphne==4.0.0
psycopg==3.1.18
boto3==1.29.4
google_api_python_client==2.107.0
google_auth_oauthlib==1.0.0
protobuf==3.20.3
pymilvus==2.3.3
gunicorn==21.2.0
supervisor==4.2.5 
sentence_transformers==2.2.2
numpy==1.21.5
anthropic
replicate
